requirements.txt
django-filter==0.15.3
djangorestframework==3.5.2
Django==1.10.3
django-extensions==1.7.4
ipdb==0.10.1
MySQL-python==1.2.5
requests==2.11.1